backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Bcrypt
PyJWT
openai-whisper
joblib
scikit-learn
numpy
pandas